api/requirements.txt
flask
flask-cors
pypdf
python-dotenv
google-generativeai
